Minor: move utils to more appropriate place
